+++ app/src/main/java/com/heyzeusv/solitaire/ui/autosizetext/DensityExt.kt
@@ -1,4 +1,4 @@
-package com.heyzeusv.solitaire.util.autosizetext
+package com.heyzeusv.solitaire.ui.autosizetext
 
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
