+++ app/src/main/java/com/heyzeusv/solitaire/ui/scoreboard/ScoreboardViewModel.kt
@@ -1,4 +1,4 @@
-package com.heyzeusv.solitaire.ui
+package com.heyzeusv.solitaire.ui.scoreboard
 
 import androidx.compose.runtime.snapshots.Snapshot
 import androidx.lifecycle.ViewModel
