+++ app/src/test/java/com/heyzeusv/solitaire/ui/CanberraViewModelTest.kt
@@ -2,6 +2,8 @@ package com.heyzeusv.solitaire.ui
 
 import com.heyzeusv.solitaire.data.Card
 import com.heyzeusv.solitaire.data.ShuffleSeed
+import com.heyzeusv.solitaire.ui.game.CanberraViewModel
+import com.heyzeusv.solitaire.ui.scoreboard.ScoreboardViewModel
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
